Agentic-ai/requirements.txt
crewai
crewai-tools
langchain
python-dotenv
requests
pinecone-client
uagents
uagents-adapter
